requirements.txt
asgiref==3.8.1
Django==5.0.6
flake8==7.1.0
mccabe==0.7.0
pillow==10.3.0
pycodestyle==2.12.0
pyflakes==3.2.0
sqlparse==0.5.0
tzdata==2024.1
